+++ java-spring/src/main/java/com/github/mangila/fibonacci/model/FibonacciOption.java
@@ -3,11 +3,11 @@ package com.github.mangila.fibonacci.model;
 import io.github.mangila.ensure4j.Ensure;
 
 public record FibonacciOption(
-        long offset,
+        int offset,
         int limit
 ) {
     public FibonacciOption {
-        Ensure.min(1, (int) offset);
+        Ensure.min(1, offset);
         Ensure.min(1, limit);
         Ensure.max(100_000, limit);
     }
