+++ java-spring/src/main/java/com/github/mangila/fibonacci/scheduler/FibonacciCompute.java
@@ -2,5 +2,5 @@ package com.github.mangila.fibonacci.scheduler;
 
 import java.math.BigInteger;
 
-public record FibonacciCompute(long id, BigInteger result) {
+public record FibonacciCompute(int id, BigInteger result) {
 }
