+++ java-spring/src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS fibonacci_results;
 CREATE TABLE IF NOT EXISTS fibonacci_results
 (
-    id     BIGINT PRIMARY KEY,
+    id     INT PRIMARY KEY,
     length INT,
     result BYTEA
 );
